Move dataclasses to models.py
